requirements.txt
requests
pillow
openai 
networkx
cloudpickle